Add base NumericData type
